Properly return references to values when unserializing
Before, this was modifying the **pointer** to the zval output variable,
which does nothing. (ZVAL_DEREF was after ZVAL_COPY)
Now, this modifies the **contents** of the zval output variable.

Handle references to an object differently from the values of that object
when serializing. See igbinary_046d.phpt in next commit.